Extract embedder and database configuration classes
As a developer, I need embedder and database configuration separated into focused files to improve maintainability and follow single responsibility principle.
