+++ mcp_server/src/config/__init__.py
@@ -1,7 +1,11 @@
 """Configuration modules for the Graphiti MCP server."""
 
+from .database_config import Neo4jConfig
+from .embedder_config import GraphitiEmbedderConfig
 from .llm_config import GraphitiLLMConfig
 
 __all__ = [
     "GraphitiLLMConfig",
+    "GraphitiEmbedderConfig",
+    "Neo4jConfig",
 ]
